Separate TermInfo out into its own file
No functional changes, just cut-and-pasted TermInfo into its own file (and made one private method on ConsolePal instead be internal).  The ConsolePal.Unix.cs file was just becoming unwieldy, with the TermInfo code being ~1000 lines.
